+++ packages/at_persistence_spec/lib/at_persistence_spec.dart
@@ -7,4 +7,5 @@ export 'package:at_persistence_spec/src/exception/exceptions.dart';
 export 'package:at_persistence_spec/src/keystore/keystore.dart';
 export 'package:at_persistence_spec/src/keystore/keystore_manager.dart';
 export 'package:at_persistence_spec/src/keystore/log_keystore.dart';
+export 'package:at_persistence_spec/src/keystore/key_pattern.dart';
 export 'package:at_persistence_spec/src/keystore/secondary_keystore.dart';